Rename fire_classifier package, update tags.
