+++ app/main.py
@@ -1,5 +1,5 @@
 from fastapi import FastAPI, File, UploadFile
-from fire_classifier.predictor import ImagePredictor
+from classifier.predictor import ImagePredictor
 from app.settings import API_SETTINGS
 
 # A: init API
@@ -15,7 +15,7 @@ predictor_config_path = API_SETTINGS["predictor_config_path"]
 predictor = ImagePredictor.init_from_config_path(predictor_config_path)
 
 
-@app.post("/classify-image/", tags=["Detect Fire"])
+@app.post("/classify-image/", tags=["Detect Fire in an Image"])
 def create_upload_file(file: UploadFile = File(...)):
     """Predicts the possibility that a RBG image contains fire."""
     return predictor.predict_from_file(file.file)
